Changed auth0 decimal data type to string. Task Member Drop Down

+++ models/Task.js
@@ -21,7 +21,7 @@ Task.init(
       defaultValue: "UNASSIGNED",
     },
     student_id: {
-      type: DataTypes.DECIMAL(32,0),
+      type: DataTypes.INTEGER,
       references: {
         model: 'student',
         key: 'id',
